Bug 1874684 - Part 8: Prefer integer types for TimeDuration. r=mgaudet
The previous changes make it possible to switch most `TimeDuration` fields to
use `int64_t` instead of `double`.

Differential Revision: https://phabricator.services.mozilla.com/D198542

UltraBlame original commit: 1e18f460744f29675581f47eadddc25dac2df516

+++ js/src/builtin/temporal/TemporalTypes.h
@@ -552,23 +552,36 @@ inline DateDuration Duration::toDateDuration() const {
 
 
 struct TimeDuration final {
-  double days = 0;
-  double hours = 0;
-  double minutes = 0;
-  double seconds = 0;
-  double milliseconds = 0;
+  
+  int64_t days = 0;
+
+  
+  int64_t hours = 0;
+
+  
+  int64_t minutes = 0;
+
+  
+  int64_t seconds = 0;
+
+  
+  int64_t milliseconds = 0;
+
+  
   double microseconds = 0;
+
+  
   double nanoseconds = 0;
 
   constexpr Duration toDuration() const {
     return {0,
             0,
             0,
-            days,
-            hours,
-            minutes,
-            seconds,
-            milliseconds,
+            double(days),
+            double(hours),
+            double(minutes),
+            double(seconds),
+            double(milliseconds),
             microseconds,
             nanoseconds};
   }
